+++ src/connection.h
@@ -49,7 +49,8 @@ parodusOnPingStatusChangeHandler on_ping_status_change;
 /*                             Function Prototypes                            */
 /*----------------------------------------------------------------------------*/
 
-int createNopollConnection(noPollCtx *);
+void set_server_list_null (server_list_t *server_list);
+int createNopollConnection(noPollCtx *, server_list_t *);
 
 /**
  * @brief Interface to terminate WebSocket client connections and clean up resources.
@@ -74,7 +75,7 @@ void set_cloud_disconnect_time(int disconnTime);
 /**
  * @brief Interface to self heal connection in progress getting stuck
  */
-void start_conn_in_progress (void);
+void start_conn_in_progress (unsigned long start_time);
 void stop_conn_in_progress (void);
 
 // To Register parodusOnPingStatusChangeHandler Callback function